+++ pystac/extensions/raster.py
@@ -625,7 +625,7 @@ class RasterBand:
         return self.properties
 
 
-class RasterExtension(PropertiesExtension, ExtensionManagementMixin[pystac.Item]):
+class RasterExtension(PropertiesExtension, ExtensionManagementMixin):
     """An abstract class that can be used to extend the properties of an
     :class:`~pystac.Item` or :class:`~pystac.Asset` with properties from
     the :stac-ext:`Raster Extension <raster>`. This class is generic over